Rename module name and file name from 'setHeroHeaderHeight' to 'heroHeader'
モジュール名とファイル名を'setHeroHeaderHeight'から'heroHeaderにリネーム。
